util/deephash: move typeIsRecursive and canMemHash to types.go
Also, rename canMemHash to typeIsMemHashable to be consistent.
There are zero changes to the semantics.
